+++ lib/sorcery/model/adapters/active_record.rb
@@ -32,7 +32,7 @@ module Sorcery
             attribute.to_s
           end
 
-          def	find_by_oauth_credentials(provider, uid)
+          def find_by_oauth_credentials(provider, uid)
             @user_config ||= ::Sorcery::Controller::Config.user_class.to_s.constantize.sorcery_config
             where(@user_config.provider_attribute_name => provider, @user_config.provider_uid_attribute_name => uid).first
           end
